// Application/Frontend/src/components/Types.ts
export interface DataBaseA {
    "field_1": string | null;
    "field_2": string | null;
    "field_3": string | null;
    "field_4": string | null;
    "field_5": string | null;
    "field_6": string | null;
    "field_7": string | null;
    "field_8": string | null;
    "field_9": string | null;
    "field_10": string | null;
    id: number;
}